+++ app/src/main/java/opensource/cached_dupe_scanner/cache/CacheDatabaseFactory.kt
@@ -24,7 +24,8 @@ val CACHE_DATABASE_MIGRATIONS: Array<Migration> = arrayOf(
     CacheMigrations.MIGRATION_15_16,
     CacheMigrations.MIGRATION_16_17,
     CacheMigrations.MIGRATION_17_18,
-    CacheMigrations.MIGRATION_18_19
+    CacheMigrations.MIGRATION_18_19,
+    CacheMigrations.MIGRATION_19_20
 )
 
 fun buildCacheDatabase(context: Context): CacheDatabase {